rename `ManagedValueState` to `InMemoryValueState`, `ManagedStateImpl` to `ManagedAggState`
