+++ src/stream/src/executor/aggregation/mod.rs
@@ -16,6 +16,8 @@ use std::any::Any;
 
 pub use agg_call::*;
 use anyhow::anyhow;
+use async_trait::async_trait;
+pub use cache::*;
 use dyn_clone::{self, DynClone};
 pub use foldable::*;
 use itertools::Itertools;
@@ -36,19 +38,27 @@ use risingwave_storage::table::streaming_table::state_table::StateTable;
 use risingwave_storage::StateStore;
 pub use row_count::*;
 pub use state_manager::*;
+pub use value::*;
 
+use self::array_agg::ManagedArrayAggState;
+use self::extreme::GenericExtremeState;
+use self::string_agg::ManagedStringAggState;
 use super::{ActorContextRef, PkIndices};
 use crate::common::{InfallibleExpression, StateTableColumnMapping};
 use crate::executor::aggregation::approx_count_distinct::StreamingApproxCountDistinct;
 use crate::executor::error::{StreamExecutorError, StreamExecutorResult};
-use crate::executor::managed_state::aggregation::ManagedStateImpl;
 use crate::executor::Executor;
 
 mod agg_call;
 mod approx_count_distinct;
+mod array_agg;
+mod cache;
+mod extreme;
 mod foldable;
 mod row_count;
 mod state_manager;
+mod string_agg;
+mod value;
 
 /// `StreamingSumAgg` sums data of the same type.
 pub type StreamingSumAgg<R, I> =
@@ -271,6 +281,7 @@ pub fn generate_agg_schema(
     Schema { fields }
 }
 
+// TODO(RC): move this to `AggStateManager::create`
 /// Create [`crate::executor::aggregation::state_manager::AggStateManager`] for `agg_calls`.
 /// For [`crate::executor::HashAggExecutor`], the group key should be provided.
 pub async fn create_agg_state_manager<S: StateStore>(
@@ -303,7 +314,7 @@ pub async fn create_agg_state_manager<S: StateStore>(
         .iter()
         .enumerate()
         .map(|(idx, agg_call)| {
-            ManagedStateImpl::create_managed_state(
+            ManagedAggState::create(
                 agg_call,
                 agg_state_tables[idx].as_ref(),
                 row_count,
@@ -369,4 +380,157 @@ pub fn for_each_agg_state_table<S: StateStore, F: Fn(&mut AggStateTable<S>)>(
         .for_each(|state_table| {
             f(state_table);
         });
+}
+
+/// Verify if the data going through the state is valid by checking if `ops.len() ==
+/// visibility.len() == data[x].len()`.
+pub fn verify_batch(
+    ops: risingwave_common::array::stream_chunk::Ops<'_>,
+    visibility: Option<&risingwave_common::buffer::Bitmap>,
+    data: &[&risingwave_common::array::ArrayImpl],
+) -> bool {
+    let mut all_lengths = vec![ops.len()];
+    if let Some(visibility) = visibility {
+        all_lengths.push(visibility.len());
+    }
+    all_lengths.extend(data.iter().map(|x| x.len()));
+    all_lengths.iter().min() == all_lengths.iter().max()
+}
+
+// TODO(RC): rename
+/// A trait over all table-structured states.
+///
+/// It is true that this interface also fits to value managed state, but we won't implement
+/// `ManagedTableState` for them. We want to reduce the overhead of `BoxedFuture`. For
+/// `ManagedValueState`, we can directly forward its async functions to `ManagedStateImpl`, instead
+/// of adding a layer of indirection caused by async traits.
+#[async_trait]
+pub trait ManagedTableState<S: StateStore>: Send + Sync + 'static {
+    async fn apply_chunk(
+        &mut self,
+        ops: Ops<'_>,
+        visibility: Option<&Bitmap>,
+        columns: &[&ArrayImpl],
+        state_table: &mut StateTable<S>,
+    ) -> StreamExecutorResult<()>;
+
+    /// Get the output of the state. Must flush before getting output.
+    async fn get_output(&mut self, state_table: &StateTable<S>) -> StreamExecutorResult<Datum>;
+}
+
+/// All managed state for aggregation. The managed state will manage the cache and integrate
+/// the state with the underlying state store. Managed states can only be evicted from outer cache
+/// when they are not dirty.
+pub enum ManagedAggState<S: StateStore> {
+    /// States as single scalar value e.g. `COUNT`, `SUM`
+    InMemoryValueState(InMemoryValueState),
+
+    /// States as table structure e.g. `MAX`, `STRING_AGG`
+    MaterializedInputState(Box<dyn ManagedTableState<S>>),
+}
+
+impl<S: StateStore> ManagedAggState<S> {
+    pub async fn apply_chunk(
+        &mut self,
+        ops: Ops<'_>,
+        visibility: Option<&Bitmap>,
+        columns: &[&ArrayImpl],
+        agg_state_table: Option<&mut AggStateTable<S>>,
+    ) -> StreamExecutorResult<()> {
+        match self {
+            Self::InMemoryValueState(state) => state.apply_chunk(ops, visibility, columns),
+            Self::MaterializedInputState(state) => {
+                state
+                    .apply_chunk(
+                        ops,
+                        visibility,
+                        columns,
+                        &mut agg_state_table
+                            .expect("managed table state must have state table")
+                            .table,
+                    )
+                    .await
+            }
+        }
+    }
+
+    /// Get the output of the state. Must flush before getting output.
+    pub async fn get_output(
+        &mut self,
+        agg_state_table: Option<&AggStateTable<S>>,
+    ) -> StreamExecutorResult<Datum> {
+        match self {
+            Self::InMemoryValueState(state) => Ok(state.get_output()),
+            Self::MaterializedInputState(state) => {
+                state
+                    .get_output(
+                        &agg_state_table
+                            .expect("managed table state must have state table")
+                            .table,
+                    )
+                    .await
+            }
+        }
+    }
+
+    /// Create a managed agg state for the given `agg_call`.
+    #[allow(clippy::too_many_arguments)]
+    pub fn create(
+        agg_call: &AggCall,
+        agg_state_table: Option<&AggStateTable<S>>,
+        row_count: usize,
+        prev_output: Option<&Datum>,
+        pk_indices: &PkIndices,
+        group_key: Option<&Row>,
+        extreme_cache_size: usize,
+        input_schema: &Schema,
+    ) -> StreamExecutorResult<Self> {
+        match agg_call.kind {
+            AggKind::Avg | AggKind::Count | AggKind::Sum | AggKind::ApproxCountDistinct => Ok(
+                Self::InMemoryValueState(InMemoryValueState::new(agg_call, prev_output.cloned())?),
+            ),
+            // optimization: use single-value state for append-only min/max
+            AggKind::Max | AggKind::Min if agg_call.append_only => Ok(Self::InMemoryValueState(
+                InMemoryValueState::new(agg_call, prev_output.cloned())?,
+            )),
+            AggKind::Max | AggKind::Min => Ok(Self::MaterializedInputState(Box::new(
+                GenericExtremeState::new(
+                    agg_call,
+                    group_key,
+                    pk_indices,
+                    agg_state_table
+                        .expect("non-append-only min/max must have state table")
+                        .mapping
+                        .clone(),
+                    row_count,
+                    extreme_cache_size,
+                    input_schema,
+                ),
+            ))),
+            AggKind::StringAgg => Ok(Self::MaterializedInputState(Box::new(
+                ManagedStringAggState::new(
+                    agg_call,
+                    group_key,
+                    pk_indices,
+                    agg_state_table
+                        .expect("string_agg must have state table")
+                        .mapping
+                        .clone(),
+                    row_count,
+                ),
+            ))),
+            AggKind::ArrayAgg => Ok(Self::MaterializedInputState(Box::new(
+                ManagedArrayAggState::new(
+                    agg_call,
+                    group_key,
+                    pk_indices,
+                    agg_state_table
+                        .expect("array_agg must have state table")
+                        .mapping
+                        .clone(),
+                    row_count,
+                ),
+            ))),
+        }
+    }
 }
