+++ src/stream/src/executor/aggregation/value.rs
@@ -21,10 +21,10 @@ use risingwave_common::types::Datum;
 use crate::executor::aggregation::{create_streaming_aggregator, AggCall, StreamingAggImpl};
 use crate::executor::error::StreamExecutorResult;
 
-/// A wrapper around [`StreamingAggImpl`], which fetches data from the state store and helps
-/// update the state. We don't use any trait to wrap around all `ManagedXxxState`, so as to reduce
-/// the overhead of creating boxed async future.
-pub struct ManagedValueState {
+/// A wrapper around [`StreamingAggImpl`], which maintains aggregation result as a value in memory.
+/// [`crate::executor::aggregation::AggStateManager`] will call `get_output` the get the result
+/// and store it into result state table.
+pub struct InMemoryValueState {
     /// Upstream column indices of agg arguments.
     arg_indices: Vec<usize>,
 
@@ -32,7 +32,7 @@ pub struct ManagedValueState {
     inner: Box<dyn StreamingAggImpl>,
 }
 
-impl ManagedValueState {
+impl InMemoryValueState {
     /// Create a single-value managed state based on `AggCall` and `Keyspace`.
     pub fn new(agg_call: &AggCall, prev_output: Option<Datum>) -> StreamExecutorResult<Self> {
         // Create the internal state based on the value we get.
@@ -93,12 +93,12 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_managed_value_state_count() {
+    async fn test_in_memory_value_state_count() {
         let agg_call = create_test_count_agg();
-        let mut managed_state = ManagedValueState::new(&agg_call, None).unwrap();
+        let mut state = InMemoryValueState::new(&agg_call, None).unwrap();
 
-        // apply a batch and get the output
-        managed_state
+        // apply a chunk
+        state
             .apply_chunk(
                 &[Op::Insert, Op::Insert, Op::Insert, Op::Insert],
                 None,
@@ -107,20 +107,20 @@ mod tests {
             .unwrap();
 
         // get output
-        let output = managed_state.get_output();
+        let output = state.get_output();
         assert_eq!(output, Some(ScalarImpl::Int64(3)));
 
         // check recovery
-        let mut managed_state = ManagedValueState::new(&agg_call, Some(output)).unwrap();
-        assert_eq!(managed_state.get_output(), Some(ScalarImpl::Int64(3)));
-        managed_state
+        let mut state = InMemoryValueState::new(&agg_call, Some(output)).unwrap();
+        assert_eq!(state.get_output(), Some(ScalarImpl::Int64(3)));
+        state
             .apply_chunk(
                 &[Op::Insert, Op::Insert, Op::Delete, Op::Insert],
                 None,
                 &[&I64Array::from_slice(&[Some(42), None, Some(2), Some(8)]).into()],
             )
             .unwrap();
-        assert_eq!(managed_state.get_output(), Some(ScalarImpl::Int64(4)));
+        assert_eq!(state.get_output(), Some(ScalarImpl::Int64(4)));
     }
 
     fn create_test_max_agg_append_only() -> AggCall {
@@ -135,12 +135,12 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_managed_value_state_append_only_max() {
+    async fn test_in_memory_value_state_append_only_max() {
         let agg_call = create_test_max_agg_append_only();
-        let mut managed_state = ManagedValueState::new(&agg_call, None).unwrap();
+        let mut state = InMemoryValueState::new(&agg_call, None).unwrap();
 
-        // apply a batch and get the output
-        managed_state
+        // apply a chunk
+        state
             .apply_chunk(
                 &[Op::Insert, Op::Insert, Op::Insert, Op::Insert, Op::Insert],
                 None,
@@ -149,6 +149,6 @@ mod tests {
             .unwrap();
 
         // get output
-        assert_eq!(managed_state.get_output(), Some(ScalarImpl::Int64(2)));
+        assert_eq!(state.get_output(), Some(ScalarImpl::Int64(2)));
     }
 }
